Feature/high res (#9)
* WIP, LoRes working again, not through GUI

* POC high res brightness control (remote, not UI)

* working HiRes controls, all fixture definitions updated

* HiRes is the definitive example now

* MIDI values need to be converted for u16 macros, too

* more readable conversion

* Animations only need to be f32

* remove unused hover label

* v0.8 supports 16-bit macro controls

+++ src/project/scene.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub enum SceneValue {
-    ControlValue(u8),
+    ControlValue(u16),
     ColourValue(Color32),
 }
